refactor: move inventory modals to features
